+++ src/main/java/calculator/StringHelper.java
@@ -1,4 +1,4 @@
-package racing;
+package calculator;
 
 import java.util.Random;
